% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{bpc_colors}
\alias{bpc_colors}
\title{BPC Corporate Colors}
\format{
A character vector with 6 HEX color codes:
\describe{
\item{1}{#6CC9D0 - Light blue}
\item{2}{#ACAD71 - Olive green}
\item{3}{#7F7B64 - Dark taupe}
\item{4}{#CF9B62 - Peach}
\item{5}{#9f2020 - Red}
\item{6}{#585458 - Dark gray}
}
}
\usage{
bpc_colors
}
\description{
The official qualitative color palette for BPC research organization.
}
\examples{
bpc_colors
scales::show_col(bpc_colors)

}
\keyword{datasets}
