% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{bpc_cols}
\alias{bpc_cols}
\title{Get BPC Color Palette}
\usage{
bpc_cols(palette = "qualitative", n = NULL, direction = 1)
}
\arguments{
\item{palette}{Name of the palette. One of: "qualitative", "sequential_1" to
"sequential_6", or "divergent_1" to "divergent_7".}

\item{n}{Number of colors desired. If NULL, uses default length for the palette.}

\item{direction}{Direction of the palette (1 = normal, -1 = reversed).}
}
\value{
A character vector of color hex codes.
}
\description{
Retrieve BPC corporate color palettes for use in plots.
}
\examples{
# Qualitative palette
bpc_cols("qualitative")

# Sequential palette
bpc_cols("sequential_1", n = 5)

# Divergent palette reversed
bpc_cols("divergent_1", direction = -1)

}
