% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{display_bpc_all}
\alias{display_bpc_all}
\title{Display All BPC Color Palettes}
\usage{
display_bpc_all(
  n = NULL,
  type = "all",
  select = NULL,
  exact.n = TRUE,
  direction = 1
)
}
\arguments{
\item{n}{Number of colors to display for each palette (if NULL, uses default)}

\item{type}{Type of palettes to display: "qual", "seq", "div", or "all"}

\item{select}{Specific palettes to display (by name)}

\item{exact.n}{If TRUE, only show palettes that can generate exactly n colors}

\item{direction}{Direction of the palette (1 = normal, -1 = reversed)}
}
\value{
A base R plot showing all available palettes
}
\description{
Create an overview of all available BPC color palettes, similar to
RColorBrewer's display.brewer.all() function.
}
\examples{
display_bpc_all()
display_bpc_all(type = "seq")
display_bpc_all(select = c("qualitative", "sequential_1", "divergent_1"))

}
