% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{scale_color_bpc_qualitative}
\alias{scale_color_bpc_qualitative}
\title{BPC Qualitative Color Scale for ggplot2}
\usage{
scale_color_bpc_qualitative(..., direction = 1)
}
\arguments{
\item{...}{Additional arguments passed to discrete_scale.}

\item{direction}{Direction of the palette (1 = normal, -1 = reversed).}
}
\value{
A ggplot2 discrete scale object.
}
\description{
Qualitative color scale using BPC corporate colors for ggplot2.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt, color = factor(cyl))) +
  geom_point() +
  scale_color_bpc_qualitative()

}
