% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{scale_color_bpc_sequential}
\alias{scale_color_bpc_sequential}
\title{BPC Sequential Color Scale for ggplot2}
\usage{
scale_color_bpc_sequential(palette = "sequential_1", direction = 1, ...)
}
\arguments{
\item{palette}{Name of the sequential palette ("sequential_1" to "sequential_6").}

\item{direction}{Direction of the palette (1 = normal, -1 = reversed).}

\item{...}{Additional arguments passed to scale_color_gradientn.}
}
\value{
A ggplot2 continuous scale object.
}
\description{
Sequential color scale using BPC corporate colors for ggplot2.
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, color = density)) +
  geom_point() +
  scale_color_bpc_sequential("sequential_1")

}
