% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{scale_fill_bpc_divergent}
\alias{scale_fill_bpc_divergent}
\title{BPC Divergent Fill Scale for ggplot2}
\usage{
scale_fill_bpc_divergent(palette = "divergent_1", direction = 1, ...)
}
\arguments{
\item{palette}{Name of the divergent palette ("divergent_1" to "divergent_7").}

\item{direction}{Direction of the palette (1 = normal, -1 = reversed).}

\item{...}{Additional arguments passed to scale_fill_gradient2.}
}
\value{
A ggplot2 continuous scale object.
}
\description{
Divergent fill scale using BPC corporate colors for ggplot2.
}
\examples{
library(ggplot2)
ggplot(msleep, aes(bodywt, brainwt, fill = sleep_total - mean(sleep_total))) +
  geom_tile() +
  scale_fill_bpc_divergent("divergent_1")

}
