% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{show_bpc_palette}
\alias{show_bpc_palette}
\title{Display BPC Palette Colors}
\usage{
show_bpc_palette(palette_name = "qualitative", n = NULL, direction = 1)
}
\arguments{
\item{palette_name}{Name of the palette to display}

\item{n}{Number of colors to show (for non-qualitative palettes)}

\item{direction}{Direction of the palette (1 = normal, -1 = reversed)}
}
\value{
A ggplot object showing the palette colors
}
\description{
Create a visualization of BPC palette colors.
}
\examples{
show_bpc_palette("qualitative")
show_bpc_palette("sequential_1", n = 9)

}
