% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{theme_bpc}
\alias{theme_bpc}
\title{BPC Custom ggplot2 Theme}
\usage{
theme_bpc()
}
\value{
A ggplot2 theme object
}
\description{
A minimal ggplot2 theme with BPC corporate styling preferences.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  theme_bpc()

}
